Prevent redundant token rate updates (#3647)
The token rates controller fetches rates for certain collections of
tokens obtained via the tokens controller. If any of these collections
change, then new rates need to be fetched.

Unfortunately, because the check to determine whether these collections
have changed is simplistic, new rates are fetched every single time the
tokens controller state is changed.

To address this, we can consider two things:

- The token rates controller only uses the addresses of tokens in these
collections to fetch rates, and it also normalizes those addresses when
updating state.
- These collections of tokens may contain duplicates.

So to reduce the number of times rates are fetched and to optimize each
request, this commit updates the tokens state callback to assemble the
aforementioned token collections into a deduplicated, normalized list of
token addresses, and then it checks to see if this list has any changes
to know whether or not to re-fetch rates.