Add on-chain validation to ERC20 wallet_watchAsset (#1745)
## Explanation

### Problem:

`wallet_watchAsset` accepts `type`, `address`, `symbol`, and `decimals`
from dapps. But it does not verify the accuracy of this information
against the chain:

- You can pass `type=ERC20`, but provide an `address` that is an EOA,
`ERC721`, or `ERC1155` contract.
- You can pass an `ERC20` contract, but with the wrong `symbol` or
`decimals`. Malicious dapps can watch scam tokens under legit symbols,
as seen here:


https://github.com/MetaMask/core/assets/3500406/55a294ab-e3af-409b-814f-e61e74653f99

### Solution:

Query the chain to get the contract type, `symbol`, and `decimals`.
Validate the request against the on-chain data. If `symbol` and
`decimals` are defined in the contract (they're optional in ERC20), then
the caller may now omit them in the request to `wallet_watchAsset` .
Otherwise they will be validated against the on-chain data.

## References

- https://github.com/MetaMask/core/issues/1207
- https://github.com/ethereum/EIPs/blob/master/EIPS/eip-747.md

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/assets-controllers`

- **BREAKING**: `TokensController.watchAsset` now performs on-chain validation of
the asset's symbol and decimals, if they're defined in the contract.  If the symbol and decimals are defined in the contract, they are no longer required to be passed to `watchAsset`

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: Alex Donesky <adonesky@gmail.com>