Add polling by `networkClientId` to `AccountTrackerController` (#3586)
## Explanation
Adds the `PollingController` to `AccountTrackerController` to help get
us ready for multichain integration. Keeps the existing API intact while
bringing in the `PollingController` methods. It's unclear if this
controller or the [Extension version of the
controller](https://github.com/MetaMask/metamask-extension/pull/21868)
will be combined and if so which pattern will used (subscription to
block tracker vs time based polling)

## References

* Fixes https://github.com/MetaMask/MetaMask-planning/issues/1029
* See https://github.com/MetaMask/metamask-extension/pull/21868

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/assets-controllers`

- **BREAKING**: `AccountTrackerController` constructor params object
requires `getCurrentChainId` and `getNetworkClientById` hooks
- **BREAKING**: `AccountTrackerController` adds a mutex to `refresh`
making it only possible for one call to be executed at time
- **ADDED**: `AccountTrackerController` exposes `accountByChainId`
property in state
- **ADDED**: `AccountTrackerController` implements `PollingController`
and can now poll by `networkClientId` via the new methods
`startPollingByNetworkClientId`, `stopPollingByPollingToken`, and
`stopPollingByPollingToken`.
- **CHANGED**: `AccountTrackerController` accepts an optional
`networkClientId` value on the `refresh` method
- **CHANGED**: `AccountTrackerController` accepts an optional
`networkClientId` value as the last parameter of the
`syncBalanceWithAddresses ` method

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: Alex Donesky <adonesky@gmail.com>