Update 'token contract address' -> 'token address'
A recent change to update TokenRatesController from the CoinGecko API to
our own API added code which refers to "token contract addresses". While
this is a more accurate name for the hexadecimal string you can use to
refer to a token, we consistently use the name "token address" in the
code rather than "token contract address", so update these references to
be consistent with existing code.