Retry failed token price updates
The `token-prices-service` will now retry token price updates if they
fail. The retry strategy is designed to let the API recover if it is
experiencing high traffic volumes.

The service now requires state in order to hold onto the retry policy.
It has been converted to a class for that reason. This required no
changes to the abstract price service type or to the
`TokenRatesController`.

Closes #2084