Enable `@typescript-eslint/no-explicit-any`
This rule is disabled in our shared ESLint rules, but we want to enable
it in this repo. Since there are many, many violations at the moment,
this commit adds `eslint-ignore` directives to each violation. This way
we can prohibit the use of `any` going forward and then correct each
existing violation as we have time.

To automatically add these ignore directives, I used
`eslint-interactive`, which is also useful for getting a quick summary
of violations. This commit also adds this tool so that we can use it
going forward.