chore(assets-controllers): Update `cockatiel` and widen range (#3682)
## Explanation

The package `cockatiel` was mistakenly pinned when it was introduced. It
has been updated to use a caret range, and the minimum has been bumped
to v3.1.2. v3.1.2 removes test files from the package, which reduces the
package size and resolves a false positive security warning from Socket.

## References

N/A

## Changelog

### `@metamask/assets-controllers`

- Changed: Update `cockatiel` to `^3.1.2`

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate