feat(assets-controllers): token price service degraded metrics
The option `onDegraded` has been added to the
`CodefiTokenPricesServiceV2` service, which lets us react to situations
where the service is still presumed to be available (i.e. there is no
outage, the circuit has not broken) but a request failed, or succeeded
too slowly. We can use this option to track instances of degraded
service in production.

The option `degradedThreshold` has also been added. This acts as a
definition of "degraded service"; if requests are slower than this
duration, we consider the service to be degraded.

Closes #3689