Use SWC with ts-node
Currently, we use `ts-node` to run scripts written in TypeScript. By
default `ts-node` has some overhead because it typechecks the code it
runs. For instance, running the `update-readme-content` script takes
around 2 seconds to run in full.

In other projects, we've solved this by integrating `ts-node` with SWC,
which is faster because it avoids typechecking, a step we don't need to
perform for scripts (as we can rely on `yarn build` in CI or TypeScript
editor integration). This gets us to around 1.4 seconds for
`update-readme-content`. This is still a lot for a simple script, but it
is an improvement.