chore(base-controller): Restore `subscrbe` property (#3698)
## Explanation

The `subscribe` property was removed in #3590 because it is no longer
used. It has been restored here so that we can release the
`@metamask/base-controller` without any breaking changes.

This property's description has been updated to clarify that it's no
longer used, and it's now marked as deprecated.

## References

Discussed in the v104 release candidate PR:
https://github.com/MetaMask/core/pull/3695#discussion_r1435225330

## Changelog

### `@metamask/base-controller`

- Changed: Deprecate the `subscribe` property of `BaseControllerV2`
- This property was used to differentiate between `BaseControllerV1` and
`BaseControllerV2` controllers. It is no longer used, so it has been
marked as deprecated.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate