feat(assets-controllers): token price service degraded metrics (#3691)
## Explanation

The option `onDegraded` has been added to the
`CodefiTokenPricesServiceV2` service, which lets us react to situations
where the service is still presumed to be available (i.e. there is no
outage, the circuit has not broken) but a request failed, or succeeded
too slowly. We can use this option to track instances of degraded
service in production.

The option `degradedThreshold` has also been added. This acts as a
definition of "degraded service"; if requests are slower than this
duration, we consider the service to be degraded.

## References

Closes #3689

## Changelog

### `@metamask/assets-controllers`

- Added: Add support to `CodefiTokenPricesServiceV2` for tracking
degraded service
- The constructor has two new options: `onDegraded` and
`degradedThreshold`. `onDegraded` is an event handler for instances of
degraded service (i.e. failed or slow requests), and `degradedThreshold`
determines how slow a request has to be before we consider service to be
degraded.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate