chore(preferences-controller): Clean up types
The `PreferencesController` types are now declared as types rather than
interfaces, in accordance with our conventions, and they are now fully
documented. The state properties have also been alphabetized so that
they are easier to maintain.

Additionally, the old "ContactEntry" type has been renamed to
"Identity", which is a more accurate name for what it is being used for
here.