Addition of "fallback" Variation to the NameController for Petnames (#3705)
## Explanation

In this PR, we introduced a "fallback" variation (denoted as "*") in the
﻿NameController specifically for the handling of petnames. In the case
of ethereum addresses, variations typically are the hexadecimal chainID,
implicitly defining petnames for a specific address within a particular
chain.

The rationale behind adding a fallback variation is to accommodate
scenarios where the petnames are not strictly tied to a particular
chain. Instead of enforcing a chain-specific petname, the fallback
variation operates on a more flexible rule: Given an Ethereum address,
if a chain-specific petname does not exist, the fallback variation is
automatically inspected to determine if a petname exists for this
general scenario.

## References

* Related to https://github.com/MetaMask/MetaMask-planning/issues/1624

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/name-controller`

- **ADDED**: Fallback variation for petnames

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate