refactor(assets-controllers): Simplify `AccountTrackerController` tests
The `AccountTrackerController` tests have been updated to remove
instances of the `PreferencesController`. Mocks are now used instead of
the real controller.

It was useful to have a reference in tests to the default
`PreferencesController` state, so I added a `getDefaultState` static
method. This was added as a method rather than an export to protect
against accidental mutations.

Relates to #3708