refactor(assets-controllers): Simplify `AccountTrackerController` tests and add `getDefaultPreferencesState` (#3736)
## Explanation

The `AccountTrackerController` tests have been updated to remove
instances of the `PreferencesController`. Mocks are now used instead of
the real controller.

It was useful to have a reference in tests to the default
`PreferencesController` state, so I added a `getDefaultPreferencesState`
function. This was added as a function to protect against accidental
mutations.

## References

Relates to #3708

## Changelog

### `@metamask/preferences-controller`

#### Added
- Add `getDefaultPreferencesState` function

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate