refactor(assets-controllers): Simplify `TokenDetectionController` tests
The `TokenDetectionController` tests have been refactored to use mocks
instead of real controller instances, and to use the `withController`
pattern to simplify controller setup and ensure that polling has
stopped after each test.

A `getDefaultTokenListState` function has been added to the
`TokenListController` to enable us to use that default state in
tests.

Relates to #3708