feat(preferences-controller): Convert to BaseControllerV2 (#3713)
## Explanation

The `PreferencesController` has been migrated to `BaseControllerV2`. As
part of this migration, the unused `name` state property has also been
removed.

## References

Closes #3708

## Changelog

### `@metamask/preferences-controller`

#### Changed
- **BREAKING:** Convert to `BaseControllerV2`
- The constructor parameters have changed; rather than accepting an
empty "config" parameter and a "state" parameter, there is now just a
single object for all constructor arguments. This object has a mandatory
`messenger` and an optional `state` property.
- Additional type exports have been added for the controller messenger
and associated types

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate