feat(queued-request-controller): Replace confirmation with switch event
When the `QueuedRequestController` requires the globally selected
network to be changed, we now emit an event rather than triggering a
confirmation. This aligns with recent designs for this use case.

The actions `NetworkController:getNetworkConfigurationByNetworkClientId`
and `ApprovalController:addRequest` were only used for triggering the
switch network confirmation, so they are no longer needed now. They
have been removed. This removes the last dependency of this controller
on the `ApprovalController`, so it has been removed as a dependency as
well.

Closes #3983