refactor(network-controller): Consolidate `setActiveNetwork` and `setProviderType`
The `setActiveNetwork` method (and action) has been updated to support
both network client IDs and built-in network types, letting us use it
for switching to any type of network.

This was done to assist with #3763

Note that the coverage has been updated, increasing coverage for
branches but reducing it for lines and statements. The reductions were
due to lines being removed; there are no lines or statements that have
become uncovered in this PR.

Closes #2020