refactor: Update `@metamask/utils` and use `createDeferredPromise` from utils (#3769)
## Explanation

The `createDeferredPromise` function that had been defined in the
`TokenRatesController` has been replaced by the function with the same
name from the `@metamask/utils` package. It behaves identically.

The `@metamask/utils` package has been updated from `^8.2.0` to `^8.3.0`
in all packages because the `createDeferredPromise` function was added
in `v8.3.0`.

## References

This is a follow-up to #3635. The function was temporarily inlined in
that PR, but the plan was always to ship it as part of utils.

This was done now because it helps somewhat with #3763, but it's not a
complete blocker.

## Changelog

This change entry applies to all packages in the diff for this PR.

#### Changed
- Bump `@metamask/utils` to `^8.3.0`

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate