fix: Handle optional 'getCurrentAccount' in incoming transaction (#3741)
## Explanation
Fixing issue `Cannot read property 'toLowerCase' of undefined` when the
`getCurrentAccount` in `getBlockNumberKey` returns `null` or
`undefined`. It fixes the symptom of the underlying issue, not the root
cause. The root cause is the `selectedAddress` passed in the
instantiation of the Transaction controller
([here](https://github.com/MetaMask/metamask-mobile/blob/main/app/core/Engine.ts#L926))
is `undefined` but per the `PreferencesController`
([here](https://github.com/MetaMask/core/blob/main/packages/preferences-controller/src/PreferencesController.ts#L46))
types it should always be a string.
Applied this solution for the v5 branch of transaction controller in
mobile:
https://github.com/MetaMask/core/commit/35c4813b757da9b61c5aef4a2f7a0ddd1db6c7d1


## Changelog



### `@metamask/transaction-controller`

- **FIXED**: Handle optional 'getCurrentAccount' before calls
`toLowerCase()` in incoming transaction.