Petnames: Add origin (#3751)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->
Adds an "origin" property to NameEntry which specificies what part of
the application was responsible for setting a Petname.

Here is the relevant enum from NameController.ts to give you a sense of
the data we are storing with each petname.
```typescript
/**
 * Enumerates the possible origins responsible for setting a petname.
 */
export enum Origin {
  // Originated from an account identity.
  ACCOUNT_IDENTITY = 'account-identity',
  // Originated from an address book entry.
  ADDRESS_BOOK = 'address-book',
  // Originated from the API (NameController.setName). This is the default.
  API = 'api',
  // Originated from the user taking action in the UI.
  UI = 'ui',
}
``` 
This allows us to keep a "paper trail" of where petname originated from.
This comes in handy, for example, if the user deletes one of their
accounts we can also delete the associated petname if we know that
petname originated from an account identity.

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->
* Related to https://github.com/MetaMask/metamask-extension/pull/21956
## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/name-controller`

- **ADDED**: Origin property on NameEntry to keep a record of the part
of the application responsible for setting that petname.

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate