chore(user-operation-controller): Remove unnecessary `peerDependency` (#3823)
## Explanation

The `@metamask/user-operation-controller` package has been updated to
remove the peer dependency `@metamask/polling-controller`. Contrary to
what the name might suggest, `@metamask/polling-controller` is not
itself a controller package, so it's not subject to our policy of making
controller dependencies peer dependencies. It's just a regular library.

## References

N/A

## Changelog

### `@metamask/user-operation-controller`
#### Removed
- Remove `@metamask/polling-controller` peer dependency
- This was mistakenly added as a peer dependency in v1. Now it's a
regular dependency.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate