Relax peer dependency constraints
When we release a new version of a controller, our Yarn constraints
enforce that we must also bump any peer dependencies that rely on that
controller to match the new version. And one of our policies is that if
we bump a peer dependency in a package, we must release a new major
version of that package. As a result, because we make changes to
controllers all the time, we also end up making major releases all the
time too. This is unnecessary if most of those changes aren't actually
breaking.

To reduce the number of major releases, this commit relaxes constraints
such that a peer dependency on a controller is only required to match
the major version of its corresponding non-peer dependency (which is
itself required to match the current version of the controller). For
instance, if the current version of `@metamask/keyring-controller` is
12.2.0, this commit would allow a peer dependency on that controller to
merely be `^12.0.0`.