Jl/transaction multichain approveTransactionsWithSameNonce networkClientId resolution (#3898)
## Explanation

Currently `approveTransactionsWithSameNonce` may need to populate the
txParams it is passed with a nonce, but the current implementation only
works correctly when calling this method while globally selected network
is on the same chainId.

To get this to work with multichain, we attempt to resolve the txParams
chainId to a networkClientId which we then use to pull a NonceTracker on
the right chain out of the trackingMap. One downside of this approach is
that it's possible to use the NonceTracker for a different network
client that's also on the same chain than the network client that this
tx is being submitted on.

I recommend reading my comment in the issue linked below for a more
thorough explanation of current state and options.

## References

See:
https://github.com/MetaMask/MetaMask-planning/issues/2023#issuecomment-1930555110

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/package-a`

- **<CATEGORY>**: Your change here
- **<CATEGORY>**: Your change here

### `@metamask/package-b`

- **<CATEGORY>**: Your change here
- **<CATEGORY>**: Your change here

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate