refactor(queued-request-controller): Migrate request count to state
The current number of requests queued by the `QueuedRequestController`
is currently stored as a private instance variable and exposed by the
`length` method and the `countChanged` event. This works, but the
conventional way of tracking and sharing state is to use controller
state. This is what controller state is for, and more broadly this is
why controllers exist: to manage state.

To better align with our conventions, the requests queue length has
been migrated into controller state. It is now represented by the state
property `queuedRequestCount`.

The old `length` method and `countChanged` events remain, but they have
been marked as deprecated. They can be removed in a future release.

Additionally, types have been added for the built-in `stateChanged`
event and `getState` action.

This was done to help unblock RPC request queue batching (#3763).
This isn't strictly required for that work, but it made it simpler to
implement.