fix: fix import ERC20 on network 1337 (#3777)
## Explanation

This PR fixes importing an ERC20 token on chainId 1337.
This issue is on extension and Mobile.

When you run a local ganache node, you use testDapp to deploy a token,
then you try to add the token to your wallet; and go back to home page
but you wont be able to see the token.

The error starts on this
[line](https://github.com/MetaMask/core/blob/f8437e9a626007de7192871452273ebd0f2edb35/packages/assets-controllers/src/TokensController.ts#L335).
It looks like it fails to pass that line to create the `newEntry` and
push it to `newTokens` array.
The error is thrown on this
[line](https://github.com/MetaMask/core/blob/f8437e9a626007de7192871452273ebd0f2edb35/packages/assets-controllers/src/token-service.ts#L136)
(See screenshot)

<img width="1510" alt="Screenshot 2024-01-12 at 18 14 32"
src="https://github.com/MetaMask/core/assets/10994169/fc3105a3-eb68-401c-865a-0119917302eb">

I see that
[tokenApi](https://github.com/consensys-vertical-apps/va-mmcx-token-api)
does not support chainId 1337 and the error is being thrown here
https://github.com/consensys-vertical-apps/va-mmcx-token-api/blob/6c1b749c33a110f9dd8004dd283f6d9f7dcb4824/src/server.ts#L522

that is because
[networkConfig](https://github.com/consensys-vertical-apps/va-mmcx-token-api/blob/6c1b749c33a110f9dd8004dd283f6d9f7dcb4824/src/constants.ts#L399)
does not contain the 1337 chainId.

After testing also with Goerli testnet, where importing ERC20 works
fine, i noticed that goerli is not supported on core so the
fetchTokenMetadata threw
[here](https://github.com/MetaMask/core/blob/f8437e9a626007de7192871452273ebd0f2edb35/packages/assets-controllers/src/token-service.ts#L82)
because `isTokenDetectionSupportedForNetwork` returned false
https://github.com/MetaMask/core/blob/f8437e9a626007de7192871452273ebd0f2edb35/packages/assets-controllers/src/assetsUtil.ts#L156.
I wasnt sure why we did not treat chainId 1337 same way as goerli.
So i removed the `|| chainId === GANACHE_CHAIN_ID` check on the utils
fct.

After the fix, i tested importing the ERC20, importing ERC721/1155,
works fine.

Lemme know if anything!  🙏 

## Video Before


https://github.com/MetaMask/core/assets/10994169/7b567366-f5df-4d1b-b61d-7d4a46711faa

## Video After


https://github.com/MetaMask/core/assets/10994169/3068a526-4c95-4c88-9c62-38acb202561f


## References

* Fixes https://github.com/MetaMask/metamask-mobile/issues/6410

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/assets-controllers`

- **<FIXED>**: Updates the fct on assetsUtils
`isTokenListSupportedForNetwork` and removed the check for
`GANACHE_CHAIN_ID`


## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate