TransactionController MultiChain Refactor (#3643)
## Explanation

Resolves: https://github.com/MetaMask/MetaMask-planning/issues/1019

## Changelog

### `@metamask/transaction-controller`

### Added
- **BREAKING:** Constructor now expects a `getNetworkClientRegistry`
callback function
- **BREAKING:** Messenger now requires `NetworkController:stateChange`
to be an allowed event
- **BREAKING:** Messenger now requires
`NetworkController:findNetworkClientByChainId` and
`NetworkController:getNetworkClientById` actions
- Adds a feature flag parameter `isMultichainEnabled ` passed via the
constructor (and defaulted to false), which when passed a truthy value
will initialize `incomingTransactionHelper`,
`pendingTransactionTracker`, `nonceTracker` per networkClientId from the
NetworkController’s `networkClientRegistry` and a
`EtherScanRemoteTransactionSource` helper per chainId in the registry.
- Adds `destroy()` method that stops/removes internal polling and
listeners
- Exports `PendingTransactionOptions` type
- Exports `TransactionControllerOptions` type
- Adds `stopAllIncomingTransactionPolling()` that stops the global
IncomingTransactionHelper and each networkClientId's
IncomingTransactionHelper


### Changed
- **BREAKING:** `approveTransactionsWithSameNonce()` now requires
`chainId` to be populated in for each TransactionParams that is passed
- `addTransaction()` now accepts optional `networkClientId` in its
options param which specifies the network client that the transaction
will be processed with during its lifecycle if the `isMultichainEnabled`
feature flag is on
- `estimateGas()` now accepts optional networkClientId as its last param
which specifies the network client that should be used to estimate the
required gas for the given transaction
- `estimateGasBuffered()` now accepts optional networkClientId as its
last param which specifies the network client that should be used to
estimate the required gas plus buffer for the given transaction
- `getNonceLock()` now accepts optional networkClientId as its last
param which specifies which the network client's nonceTracker should be
used to determine the next nonce.
- When used with the `enableMultichain` feature flag on and with
networkClientId specified, this method will also restrict acquiring the
next nonce by chainId, i.e. if this method is called with two different
networkClientIds on the same chainId, only the first call will return
immediately with a lock from its respective nonceTracker with the second
call being blocked until the first caller releases its lock
- Approval of transactions previously used the chainId from the global
provider state as the source of truth. Instead it will now use the
chainId from the TransactionMeta object. **Not sure we really need to
put this in the changelog since all tx will have chainId populated**
- `EtherscanRemoteTransactionSource` now enforces a 5 second delay
between requests to avoid rate limiting
- `TransactionMeta` type now specifies an optional `networkClientId`
field
- `startIncomingTransactionPolling()` now accepts an optional array of
networkClientIds. When provided, the IncomingTransactionHelpers for
those networkClientIds will be started. If empty or not provided, the
global IncomingTransactionHelper will be started.
- `stopIncomingTransactionPolling()` now accepts an optional array of
networkClientIds. When provided, the IncomingTransactionHelpers for
those networkClientIds will be stoppped. If empty or not provided, the
global IncomingTransactionHelper will be stopped.

### Removed


## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: Jiexi Luan <jiexiluan@gmail.com>
Co-authored-by: Alex Donesky <adonesky@gmail.com>
Co-authored-by: Matthew Walsh <matthew.walsh@consensys.net>
Co-authored-by: Elliot Winkler <elliot.winkler@gmail.com>